refactor: extract makeMarshallToStorage
